ai/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
requests
python-dotenv
chromadb
sentence-transformers
langchain-chroma
langchain-huggingface
pysqlite3-binary
